Add shell e2e for the .NET dd_table_for_deal CLI.
Drive `dotnet run` from bash/PowerShell and run those scripts in Linux, macOS, and Windows CI.
